+++ qutebrowser/browser/commands.py
@@ -1355,8 +1355,8 @@ class CommandDispatcher:
         if dest is None:
             suggested_fn = self._current_title() + ".mht"
             suggested_fn = utils.sanitize_filename(suggested_fn)
-            filename, q = downloads.ask_for_filename(suggested_fn, parent=tab,
-                                                     url=tab.url())
+            filename, q = downloads.ask_for_filename_async(
+                suggested_fn, parent=tab, url=tab.url())
             if filename is not None:
                 mhtml.start_download_checked(filename, tab=tab)
             else:
